feat: added retry for non profit blockchain transactions
