+++ spec/jobs/person_payments/retry_blockchain_transactions_job_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe PersonPayments::RetryBlockchainTransactionsJob, type: :job do
     subject(:perform_job) { described_class.perform_now }
 
     before do
-      allow(PersonPayments::BlockchainTransactions::Cause::UpdateProcessingTransactions).to receive(:call)
+      allow(PersonPayments::BlockchainTransactions::UpdateProcessingTransactions).to receive(:call)
       allow(PersonPayments::BlockchainTransactions::Cause::UpdateApiOnlyTransactions).to receive(:call)
       allow(PersonPayments::BlockchainTransactions::Cause::UpdateFailedTransactions).to receive(:call)
       allow(PersonPayments::BlockchainTransactions::NonProfit::UpdateFailedTransactions).to receive(:call)
@@ -13,8 +13,8 @@ RSpec.describe PersonPayments::RetryBlockchainTransactionsJob, type: :job do
     end
 
     it 'calls the retries commands' do
+      expect(PersonPayments::BlockchainTransactions::UpdateProcessingTransactions).to have_received(:call)
       expect(PersonPayments::BlockchainTransactions::Cause::UpdateApiOnlyTransactions).to have_received(:call)
-      expect(PersonPayments::BlockchainTransactions::Cause::UpdateProcessingTransactions).to have_received(:call)
       expect(PersonPayments::BlockchainTransactions::Cause::UpdateFailedTransactions).to have_received(:call)
       expect(PersonPayments::BlockchainTransactions::NonProfit::UpdateFailedTransactions).to have_received(:call)
     end
